translation-worker/requirements.txt
redis==5.0.1
rq==1.15.1
PyPDF2==3.0.1
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.3.2
pdf2image==1.17.0 